refactor: move responseDTOs to new package
